Clean up test infrastructure
